# frontend/generated/backend-openapi.json
{
  "components": {
    "schemas": {
...
      "ExcludeProviderRequest": {
        "properties": {
          "runner_id": {
            "title": "Runner Id",
            "type": "string"
          }
        },
        "required": [
          "runner_id"
        ],
        "title": "ExcludeProviderRequest",
        "type": "object"
      },
...
      "HTTPErrorResponse": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "HTTPErrorResponse",
        "type": "object"
      },
...
      "SelectProviderRequest": {
        "properties": {
          "runner_id": {
            "title": "Runner Id",
            "type": "string"
          }
        },
        "required": [
          "runner_id"
        ],
        "title": "SelectProviderRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LTX-2 Video Generation Server",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/providers": {
      "get": {
        "description": "List discovered providers with selection/exclusion state.",
        "operationId": "get_providers_api_providers_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Providers Api Providers Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Get Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/api/providers/discover": {
      "post": {
        "description": "Trigger a fresh discovery sweep.",
        "operationId": "discover_providers_api_providers_discover_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Discover Providers Api Providers Discover Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Discover Providers",
        "tags": [
          "providers"
        ]
      }
    },
    "/api/providers/exclude": {
      "post": {
        "description": "Toggle exclusion for a provider.",
        "operationId": "exclude_provider_api_providers_exclude_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExcludeProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Exclude Provider Api Providers Exclude Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Exclude Provider",
        "tags": [
          "providers"
        ]
      }
    },
    "/api/providers/select": {
      "post": {
        "description": "Select a provider for inference.",
        "operationId": "select_provider_api_providers_select_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Select Provider Api Providers Select Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "4XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Client Error"
          },
          "5XX": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPErrorResponse"
                }
              }
            },
            "description": "Server Error"
          }
        },
        "summary": "Select Provider",
        "tags": [
          "providers"
        ]
      }
    },
...
  }
}